+++ lib/lightning_web/components/loaders.ex
@@ -14,7 +14,7 @@ defmodule LightningWeb.Components.Loaders do
     ~H"""
     <span class="relative inline-flex">
       <div class="inline-flex">
-        <%= render_slot(@inner_block) %>
+        {render_slot(@inner_block)}
       </div>
       <span class="flex absolute h-3 w-3 right-0 -mr-5">
         <span class="animate-ping absolute inline-flex h-full w-full rounded-full bg-primary-400 opacity-75">
@@ -39,7 +39,7 @@ defmodule LightningWeb.Components.Loaders do
             ring-slate-900/10 dark:ring-slate-200/20"
         disabled=""
       >
-        <%= render_slot(@inner_block) %>
+        {render_slot(@inner_block)}
       </button>
       <span class="flex absolute h-3 w-3 top-0 right-0 -mt-1 -mr-1">
         <span class="animate-ping absolute inline-flex h-full w-full rounded-full bg-primary-400 opacity-75">