+++ lib/lightning_web/live/components/form.ex
@@ -15,7 +15,7 @@ defmodule LightningWeb.Components.Form do
       px-4
       border
       border-transparent
-      shadow-sm
+      shadow-xs
       text-sm
       font-medium
       rounded-md
@@ -35,7 +35,7 @@ defmodule LightningWeb.Components.Form do
 
     ~H"""
     <button type="submit" class={@class} {@rest}>
-      <%= render_slot(@inner_block) %>
+      {render_slot(@inner_block)}
     </button>
     """
   end
@@ -72,18 +72,18 @@ defmodule LightningWeb.Components.Form do
     <div class="flex">
       <div class="shrink">
         <%= if @label do %>
-          <%= PhoenixHTMLHelpers.Form.label(@form, @field, @label,
+          {PhoenixHTMLHelpers.Form.label(@form, @field, @label,
             class: @label_classes
-          ) %>
+          )}
         <% else %>
-          <%= PhoenixHTMLHelpers.Form.label(@form, @field, class: @label_classes) %>
+          {PhoenixHTMLHelpers.Form.label(@form, @field, class: @label_classes)}
         <% end %>
       </div>
       <div class="grow text-right">
         <.old_error field={@form[@field]} />
       </div>
     </div>
-    <%= PhoenixHTMLHelpers.Form.textarea(@form, @field, @opts) %>
+    {PhoenixHTMLHelpers.Form.textarea(@form, @field, @opts)}
     """
   end
 
@@ -116,7 +116,7 @@ defmodule LightningWeb.Components.Form do
       focus:ring-primary-500
       focus:border-primary-500
       block w-full
-      shadow-sm
+      shadow-xs
       sm:text-sm
       border-secondary-300
       rounded-md
@@ -137,18 +137,18 @@ defmodule LightningWeb.Components.Form do
 
     ~H"""
     <%= if @label do %>
-      <%= PhoenixHTMLHelpers.Form.label(@form, @id, @label, class: @label_classes) %>
+      {PhoenixHTMLHelpers.Form.label(@form, @id, @label, class: @label_classes)}
     <% else %>
-      <%= PhoenixHTMLHelpers.Form.label(@form, @id, class: @label_classes) %>
+      {PhoenixHTMLHelpers.Form.label(@form, @id, class: @label_classes)}
     <% end %>
-    <%= if assigns[:inner_block], do: render_slot(@inner_block) %>
+    {if assigns[:inner_block], do: render_slot(@inner_block)}
 
     <LightningWeb.CoreComponents.old_error field={@form[@id]} />
-    <%= PhoenixHTMLHelpers.Form.password_input(
+    {PhoenixHTMLHelpers.Form.password_input(
       @form,
       @id,
       @opts ++ [class: @input_classes, required: @required, value: @value]
-    ) %>
+    )}
     """
   end
 
@@ -171,7 +171,7 @@ defmodule LightningWeb.Components.Form do
       focus:ring-primary-500
       focus:border-primary-500
       block w-full
-      shadow-sm
+      shadow-xs
       sm:text-sm
       border-secondary-300
       rounded-md
@@ -186,12 +186,12 @@ defmodule LightningWeb.Components.Form do
       |> assign_new(:required, fn -> false end)
 
     ~H"""
-    <%= PhoenixHTMLHelpers.Form.label(@form, @id, class: @label_classes) %>
+    {PhoenixHTMLHelpers.Form.label(@form, @id, class: @label_classes)}
     <LightningWeb.CoreComponents.old_error field={@form[@id]} />
-    <%= PhoenixHTMLHelpers.Form.email_input(@form, @id,
+    {PhoenixHTMLHelpers.Form.email_input(@form, @id,
       class: @input_classes,
       required: @required
-    ) %>
+    )}
     """
   end
 
@@ -248,7 +248,7 @@ defmodule LightningWeb.Components.Form do
       border-0
       py-1.5
       text-gray-900
-      shadow-sm
+      shadow-xs
       ring-1
       ring-gray-300
       placeholder:text-gray-400
@@ -274,16 +274,16 @@ defmodule LightningWeb.Components.Form do
 
     ~H"""
     <%= if @label do %>
-      <%= PhoenixHTMLHelpers.Form.label(@form, @field, @label, class: @label_classes) %>
+      {PhoenixHTMLHelpers.Form.label(@form, @field, @label, class: @label_classes)}
     <% else %>
-      <%= PhoenixHTMLHelpers.Form.label(@form, @field, class: @label_classes) %>
+      {PhoenixHTMLHelpers.Form.label(@form, @field, class: @label_classes)}
     <% end %>
-    <%= if assigns[:inner_block], do: render_slot(@inner_block) %>
-    <%= PhoenixHTMLHelpers.Form.text_input(
+    {if assigns[:inner_block], do: render_slot(@inner_block)}
+    {PhoenixHTMLHelpers.Form.text_input(
       @form,
       @field,
       @opts ++ [class: @input_classes, required: @required, disabled: @disabled]
-    ) %>
+    )}
     <.old_error field={@form[@field]} />
     """
   end
@@ -319,7 +319,7 @@ defmodule LightningWeb.Components.Form do
       focus:border-primary-500
       block
       w-full
-      shadow-sm
+      shadow-xs
       text-sm
       border-secondary-300
       rounded-md
@@ -342,12 +342,12 @@ defmodule LightningWeb.Components.Form do
     ~H"""
     <div class="flex items-start">
       <div class="flex items-center h-5">
-        <%= PhoenixHTMLHelpers.Form.checkbox(@form, @field, @checkbox_opts) %>
+        {PhoenixHTMLHelpers.Form.checkbox(@form, @field, @checkbox_opts)}
       </div>
       <div class="ml-3 text-sm">
         <LightningWeb.CoreComponents.old_error field={@form[@field]} />
-        <%= PhoenixHTMLHelpers.Form.label(@form, @label || @field, @label_opts) %>
-        <%= render_slot(@inner_block) %>
+        {PhoenixHTMLHelpers.Form.label(@form, @label || @field, @label_opts)}
+        {render_slot(@inner_block)}
       </div>
     </div>
     """
@@ -383,7 +383,7 @@ defmodule LightningWeb.Components.Form do
       <div class="flex flex-row items-end">
         <%= PhoenixHTMLHelpers.Form.label(@form, @field, @opts) do %>
           <div class="flex items-center">
-            <%= @title %>
+            {@title}
             <img :if={@logo} src={@logo} class="w-3 h-3 ml-1" />
           </div>
         <% end %>
@@ -395,7 +395,7 @@ defmodule LightningWeb.Components.Form do
     <% else %>
       <%= PhoenixHTMLHelpers.Form.label(@form, @field, @opts) do %>
         <div class="flex items-center">
-          <%= @title %>
+          {@title}
           <img :if={@logo} src={@logo} class="w-3 h-3 ml-1" />
         </div>
       <% end %>
@@ -416,13 +416,12 @@ defmodule LightningWeb.Components.Form do
       w-full
       rounded-md
       border-secondary-300
-      shadow-sm
+      shadow-xs
       text-md
       font-medium
       focus:border-primary-300
       focus:ring
-      focus:ring-primary-200
-      focus:ring-opacity-50
+      focus:ring-primary-200/50
       disabled:cursor-not-allowed
     ]
 
@@ -433,7 +432,7 @@ defmodule LightningWeb.Components.Form do
     assigns = assign(assigns, opts: opts)
 
     ~H"""
-    <%= PhoenixHTMLHelpers.Form.select(@form, @name, @values, @opts) %>
+    {PhoenixHTMLHelpers.Form.select(@form, @name, @values, @opts)}
     """
   end
 
@@ -444,11 +443,10 @@ defmodule LightningWeb.Components.Form do
       rounded-md
       border-secondary-300
       sm:text-sm
-      shadow-sm
+      shadow-xs
       focus:border-primary-300
       focus:ring
-      focus:ring-primary-200
-      focus:ring-opacity-50
+      focus:ring-primary-200/50
       disabled:cursor-not-allowed
     ]
 
@@ -458,7 +456,7 @@ defmodule LightningWeb.Components.Form do
 
     ~H"""
     <select {@opts}>
-      <%= render_slot(@inner_block) %>
+      {render_slot(@inner_block)}
     </select>
     """
   end