+++ lib/lightning_web/live/credential_live/json_schema_body_component.ex
@@ -29,14 +29,14 @@ defmodule LightningWeb.CredentialLive.JsonSchemaBodyComponent do
       )
 
     ~H"""
-    <%= render_slot(
+    {render_slot(
       @inner_block,
       {Phoenix.LiveView.TagEngine.component(
          &inner/1,
          [form: @form, schema_changeset: @schema_changeset, schema: @schema],
          {__ENV__.module, __ENV__.function, __ENV__.file, __ENV__.line}
        ), @valid?}
-    ) %>
+    )}
     """
   end