+++ lib/lightning_web/live/project_live/invite_collaborator_component.ex
@@ -80,7 +80,7 @@ defmodule LightningWeb.ProjectLive.InviteCollaboratorComponent do
           &LightningWeb.CoreComponents.translate_error(&1)
         )
     }>
-      <%= msg %>
+      {msg}
     </.error>
     """
   end