+++ test/lightning_web/live/project_live_test.exs
@@ -303,7 +303,7 @@ defmodule LightningWeb.ProjectLiveTest do
           Routes.project_index_path(conn, :index)
         )
 
-      assert has_element?(index_live, "#project-#{project.id}")
+      assert has_element?(index_live, "#delete-#{project.id}")
     end
 
     test "allows a superuser to cancel scheduled deletion on a project", %{
@@ -313,10 +313,10 @@ defmodule LightningWeb.ProjectLiveTest do
         project_fixture(scheduled_deletion: Timex.now() |> Timex.shift(days: 7))
 
       {:ok, index_live, _html} =
-        live(conn, Routes.project_index_path(conn, :index))
+        live(conn, Routes.project_index_path(conn, :index), on_error: :raise)
 
       assert index_live
-             |> element("#project-#{project.id} a", "Cancel deletion")
+             |> element("#cancel-deletion-#{project.id}", "Cancel deletion")
              |> render_click() =~ "Project deletion canceled"
     end
 
@@ -328,11 +328,11 @@ defmodule LightningWeb.ProjectLiveTest do
         project_fixture(scheduled_deletion: Timex.now() |> Timex.shift(days: 7))
 
       {:ok, index_live, _html} =
-        live(conn, Routes.project_index_path(conn, :index))
+        live(conn, Routes.project_index_path(conn, :index), on_error: :raise)
 
       {:ok, form_live, _html} =
         index_live
-        |> element("#project-#{project.id} a", "Delete now")
+        |> element("#delete-now-#{project.id}", "Delete now")
         |> render_click()
         |> follow_redirect(
           conn,
@@ -362,7 +362,8 @@ defmodule LightningWeb.ProjectLiveTest do
       project =
         insert(:project, project_users: [%{role: :owner, user_id: user1.id}])
 
-      {:ok, view, _html} = live(conn, ~p"/settings/projects/#{project.id}")
+      {:ok, view, _html} =
+        live(conn, ~p"/settings/projects/#{project.id}", on_error: :raise)
 
       view
       |> form("#project-form",
@@ -493,7 +494,8 @@ defmodule LightningWeb.ProjectLiveTest do
           project_users: [%{user_id: another_user.id}]
         )
 
-      {:ok, view, _html} = live(conn, ~p"/projects/#{project_1}/w")
+      {:ok, view, _html} =
+        live(conn, ~p"/projects/#{project_1}/w", on_error: :raise)
 
       assert view
              |> element(
@@ -516,7 +518,8 @@ defmodule LightningWeb.ProjectLiveTest do
              )
              |> has_element?()
 
-      {:ok, view, html} = live(conn, ~p"/projects/#{project_1}/w")
+      {:ok, view, html} =
+        live(conn, ~p"/projects/#{project_1}/w", on_error: :raise)
 
       assert html =~ project_1.name
 
@@ -532,7 +535,8 @@ defmodule LightningWeb.ProjectLiveTest do
              |> element("#option-#{project_3.id}")
              |> has_element?()
 
-      {:ok, view, html} = live(conn, ~p"/projects/#{project_2}/w")
+      {:ok, view, html} =
+        live(conn, ~p"/projects/#{project_2}/w", on_error: :raise)
 
       assert html =~ project_2.name
 
@@ -548,7 +552,7 @@ defmodule LightningWeb.ProjectLiveTest do
              |> element("#option-#{project_3.id}")
              |> has_element?()
 
-      assert live(conn, ~p"/projects/#{project_3}/w") ==
+      assert live(conn, ~p"/projects/#{project_3}/w", on_error: :raise) ==
                {:error,
                 {:redirect, %{flash: %{"nav" => :not_found}, to: "/projects"}}}
     end
@@ -562,7 +566,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, _view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Project settings"
@@ -580,7 +585,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, _view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Collaborator"
@@ -625,7 +631,8 @@ defmodule LightningWeb.ProjectLiveTest do
         live(
           conn,
           Routes.project_project_settings_path(conn, :index, project.id) <>
-            "#credentials"
+            "#credentials",
+          on_error: :raise
         )
 
       assert html =~ "Name"
@@ -658,7 +665,8 @@ defmodule LightningWeb.ProjectLiveTest do
           live(
             conn,
             Routes.project_project_settings_path(conn, :index, project.id) <>
-              "#credentials"
+              "#credentials",
+            on_error: :raise
           )
 
         credential_name = Lightning.Name.generate()
@@ -705,7 +713,8 @@ defmodule LightningWeb.ProjectLiveTest do
         live(
           conn,
           Routes.project_project_settings_path(conn, :index, project.id) <>
-            "#credentials"
+            "#credentials",
+          on_error: :raise
         )
 
       credential_name = "My Credential"
@@ -753,7 +762,8 @@ defmodule LightningWeb.ProjectLiveTest do
         live(
           conn,
           Routes.project_project_settings_path(conn, :index, project.id) <>
-            "#credentials"
+            "#credentials",
+          on_error: :raise
         )
 
       refute html =~ credential_name
@@ -777,7 +787,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Project settings"
@@ -821,7 +832,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Project settings"
@@ -855,7 +867,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Project settings"
@@ -881,19 +894,18 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Project settings"
 
       assert view
-             |> has_element?(
-               "input[disabled='disabled'][id='project-settings-form_name']"
-             )
+             |> has_element?("input[disabled='disabled'][name='project[name]']")
 
       assert view
              |> has_element?(
-               "textarea[disabled='disabled'][id='project-settings-form_description']"
+               "textarea[disabled='disabled'][name='project[description]']"
              )
 
       assert view |> has_element?("button[disabled][type=submit]")
@@ -927,7 +939,8 @@ defmodule LightningWeb.ProjectLiveTest do
         live(
           conn,
           Routes.project_project_settings_path(conn, :index, project.id) <>
-            "#collaboration"
+            "#collaboration",
+          on_error: :raise
         )
 
       authenticated_user_project_user =
@@ -963,7 +976,7 @@ defmodule LightningWeb.ProjectLiveTest do
              )
 
       view
-      |> element("#flash")
+      |> element("[data-flash-kind='info']")
       |> render_hook("lv:clear-flash")
 
       form_id = "#digest-#{authenticated_user_project_user.id}"
@@ -997,7 +1010,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, view, html} =
           live(
             conn,
-            Routes.project_project_settings_path(conn, :index, project.id)
+            Routes.project_project_settings_path(conn, :index, project.id),
+            on_error: :raise
           )
 
         assert has_element?(view, "#security-tab")
@@ -1018,7 +1032,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Project settings"
@@ -1044,7 +1059,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, view, html} =
           live(
             conn,
-            Routes.project_project_settings_path(conn, :index, project.id)
+            Routes.project_project_settings_path(conn, :index, project.id),
+            on_error: :raise
           )
 
         assert html =~ "Project settings"
@@ -1075,7 +1091,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, _view, html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert html =~ "Project settings"
@@ -1087,7 +1104,12 @@ defmodule LightningWeb.ProjectLiveTest do
         assert {:error, {:redirect, %{to: "/mfa_required"}}} =
                  live(
                    conn,
-                   Routes.project_project_settings_path(conn, :index, project.id)
+                   Routes.project_project_settings_path(
+                     conn,
+                     :index,
+                     project.id
+                   ),
+                   on_error: :raise
                  )
       end)
     end
@@ -1101,7 +1123,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, _view, html} =
           live(
             conn,
-            Routes.project_project_settings_path(conn, :index, project.id)
+            Routes.project_project_settings_path(conn, :index, project.id),
+            on_error: :raise
           )
 
         for auth_method <- auth_methods do
@@ -1123,7 +1146,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, view, _html} =
           live(
             conn,
-            settings_path
+            settings_path,
+            on_error: :raise
           )
 
         assert view |> element("button#add_new_auth_method") |> has_element?()
@@ -1171,7 +1195,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, _view, html} =
           live(
             conn,
-            settings_path
+            settings_path,
+            on_error: :raise
           )
 
         assert html =~ credential_name
@@ -1181,7 +1206,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, view, _html} =
           live(
             conn,
-            settings_path
+            settings_path,
+            on_error: :raise
           )
 
         assert view
@@ -1211,7 +1237,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, _html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert view
@@ -1240,7 +1267,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, view, _html} =
           live(
             conn,
-            settings_path
+            settings_path,
+            on_error: :raise
           )
 
         assert view
@@ -1282,7 +1310,8 @@ defmodule LightningWeb.ProjectLiveTest do
         {:ok, view, _html} =
           live(
             conn,
-            settings_path
+            settings_path,
+            on_error: :raise
           )
 
         assert view
@@ -1320,7 +1349,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, _html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert view
@@ -1359,7 +1389,8 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, _html} =
         live(
           conn,
-          Routes.project_project_settings_path(conn, :index, project.id)
+          Routes.project_project_settings_path(conn, :index, project.id),
+          on_error: :raise
         )
 
       assert view
@@ -1524,7 +1555,7 @@ defmodule LightningWeb.ProjectLiveTest do
           attrs: %{text: banner_message},
           function: fn assigns ->
             ~H"""
-            <%= @text %>
+            {@text}
             """
           end
         }
@@ -1629,15 +1660,15 @@ defmodule LightningWeb.ProjectLiveTest do
           settings_path
         )
 
-      refute view
-             |> element("a#delete_auth_method_link_#{auth_method.id}")
-             |> has_element?()
+      assert view
+             |> has_element?(
+               "a#delete_auth_method_link_#{auth_method.id}.cursor-not-allowed"
+             )
 
       modal_id = "delete_auth_#{auth_method.id}_modal"
 
       refute view
-             |> element("#delete_auth_method_#{modal_id}_#{auth_method.id}")
-             |> has_element?()
+             |> has_element?("#delete_auth_method_#{modal_id}_#{auth_method.id}")
     end
   end
 
@@ -1913,7 +1944,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
       html =
         view
-        |> element("select#retention-settings-form_history_retention_period")
+        |> element("select[name='project[history_retention_period]']")
         |> render()
 
       for option <- expected_options do
@@ -1948,7 +1979,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
       assert has_element?(
                view,
-               "#retention-settings-form_history_retention_period:disabled"
+               "select[name='project[history_retention_period]']:disabled"
              )
     end
 
@@ -1969,7 +2000,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
       assert has_element?(
                view,
-               "#retention-settings-form_dataclip_retention_period:disabled"
+               "select[name='project[dataclip_retention_period]']:disabled"
              )
 
       view
@@ -1982,12 +2013,12 @@ defmodule LightningWeb.ProjectLiveTest do
 
       refute has_element?(
                view,
-               "#retention-settings-form_dataclip_retention_period:disabled"
+               "select[name='project[dataclip_retention_period]']:disabled"
              )
 
       assert has_element?(
                view,
-               "#retention-settings-form_dataclip_retention_period"
+               "select[name='project[dataclip_retention_period]']"
              )
     end
 
@@ -2006,7 +2037,7 @@ defmodule LightningWeb.ProjectLiveTest do
       selected_dataclip_option =
         element(
           view,
-          "#retention-settings-form_dataclip_retention_period option[selected]"
+          "select[name='project[dataclip_retention_period]'] option[selected]"
         )
 
       # nothing has been selected for the dataclip period
@@ -2024,7 +2055,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
       refute has_element?(
                view,
-               "#retention-settings-form_dataclip_retention_period:disabled"
+               "#retention-settings-form select[name='project[dataclip_retention_period]']:disabled"
              )
 
       # 7 Days has been selected for the dataclip period
@@ -2041,7 +2072,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
       assert has_element?(
                view,
-               "#retention-settings-form_dataclip_retention_period:disabled"
+               "#retention-settings-form select[name='project[dataclip_retention_period]']:disabled"
              )
 
       # 7 days gets cleared. Nothing is now selected
@@ -3918,7 +3949,7 @@ defmodule LightningWeb.ProjectLiveTest do
           {:error, :disabled,
            %{
              function: fn assigns ->
-               ~H"<p>I am an error message that says: <%= @error %></p>"
+               ~H"<p>I am an error message that says: {@error}</p>"
              end,
              attrs: %{error: error_msg}
            }}
@@ -4845,7 +4876,7 @@ defmodule LightningWeb.ProjectLiveTest do
           {:error, :disabled,
            %{
              function: fn assigns ->
-               ~H"<p>I am an error message that says: <%= @error %></p>"
+               ~H"<p>I am an error message that says: {@error}</p>"
              end,
              attrs: %{error: error_msg}
            }}
@@ -4969,7 +5000,7 @@ defmodule LightningWeb.ProjectLiveTest do
             {:error, :disabled,
              %{
                function: fn assigns ->
-                 ~H"<p>I am an error message that says: <%= @error %></p>"
+                 ~H"<p>I am an error message that says: {@error}</p>"
                end,
                attrs: %{error: error_msg}
              }}