+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to
 
 - Added alphabetical sort to credential and Oauth2 client lists.
   [#2994](https://github.com/OpenFn/lightning/issues/2994)
+- Update Phoenix LiveView to 1.0.5, Petal Components to 2.9.0 and TailwindCSS to
+  v4 [#3033](https://github.com/OpenFn/lightning/pull/3033)
 
 ### Fixed